Add BoundingBox

+++ pizzaserver-commons/src/main/java/io/github/willqi/pizzaserver/commons/utils/Vector2.java
@@ -75,29 +75,29 @@ public class Vector2 {
         return new Vector2(this.getX() / x, this.getY() / y);
     }
 
-    public double distanceBetween(Vector2i vector2i) {
-        return Math.sqrt(Math.pow(this.getX() - vector2i.getX(), 2) + Math.pow(this.getY() - vector2i.getY(), 2));
+    public float distanceBetween(Vector2i vector2i) {
+        return (float) Math.sqrt(Math.pow(this.getX() - vector2i.getX(), 2) + Math.pow(this.getY() - vector2i.getY(), 2));
     }
 
-    public double distanceBetween(Vector2 vector2) {
-        return Math.sqrt(Math.pow(this.getX() - vector2.getX(), 2) + Math.pow(this.getY() - vector2.getY(), 2));
+    public float distanceBetween(Vector2 vector2) {
+        return (float) Math.sqrt(Math.pow(this.getX() - vector2.getX(), 2) + Math.pow(this.getY() - vector2.getY(), 2));
     }
 
-    public double getLength() {
-        return Math.sqrt(Math.pow(this.getX(), 2) + Math.pow(this.getY(), 2));
+    public float getLength() {
+        return (float) Math.sqrt(Math.pow(this.getX(), 2) + Math.pow(this.getY(), 2));
     }
 
-    public double dot(Vector2 vector2) {
+    public float dot(Vector2 vector2) {
         return (this.getX() * vector2.getX()) + (this.getY() * vector2.getY());
     }
 
-    public double dot(Vector2i vector2i) {
+    public float dot(Vector2i vector2i) {
         return (this.getX() * vector2i.getX()) + (this.getY() * vector2i.getY());
     }
 
     public Vector2 normalize() {
-        double length = this.getLength();
-        return new Vector2((float) (this.getX() / length), (float) (this.getY() / length));
+        float length = this.getLength();
+        return new Vector2(this.getX() / length, this.getY() / length);
     }
 
     public Vector2i toVector2i() {
